Rename IngestRecordsLocalStackIT
