+++ java/ingest/ingest-runner/src/test/java/sleeper/ingest/IngestRecordsLocalStackITBase.java
@@ -34,7 +34,7 @@ import static sleeper.configuration.testutils.LocalStackAwsV1ClientHelper.buildA
 import static sleeper.utils.HadoopConfigurationLocalStackUtils.getHadoopConfiguration;
 
 @Testcontainers
-public class IngestRecordsS3ITBase extends IngestRecordsTestBase {
+public class IngestRecordsLocalStackITBase extends IngestRecordsTestBase {
     @Container
     public static LocalStackContainer localStackContainer = new LocalStackContainer(DockerImageName.parse(CommonTestConstants.LOCALSTACK_DOCKER_IMAGE))
             .withServices(LocalStackContainer.Service.S3, LocalStackContainer.Service.DYNAMODB);
